move getrandomid to helpers directory
